Create .gitattributes
This change to ignore `Jupyter notebook` from being identified as the repo's language (since notebook files are way more lengthy comparing to the Python files, and fool the GitHub repository's language analyser). This might be helpful when people exploring for solutions by GitHub search function, and also simply because the repo's language should be `Python`.